Extract newton world interface + update particles in the real world
